# feat(common): support fixed int enum