--- a/src/AutoRest.CSharp/Common/Input/CadlInputEnumTypeConverter.cs
+++ b/src/AutoRest.CSharp/Common/Input/CadlInputEnumTypeConverter.cs
@@ -71,12 +71,56 @@ namespace AutoRest.CSharp.Common.Input
                 throw new JsonException("Enum must have at least one value");
             }
 
-            var enumType = new InputEnumType(name, ns, accessibility, description, usage, valueType ?? InputPrimitiveType.Int32, allowedValues, isExtendable);
+            valueType ??= InputPrimitiveType.Int32;
+            var enumType = new InputEnumType(name, ns, accessibility, description, usage, valueType, normalizeValues(allowedValues, valueType), isExtendable);
             if (id != null)
             {
                 resolver.AddReference(id, enumType);
             }
             return enumType;
         }
+
+        private static IReadOnlyList<InputEnumTypeValue> normalizeValues(IReadOnlyList<InputEnumTypeValue> allowedValues, InputPrimitiveType valueType)
+        {
+            var concreteValues = new List<InputEnumTypeValue>(allowedValues.Count);
+
+            switch (valueType.Kind)
+            {
+                case InputTypeKind.String:
+                    foreach (var value in allowedValues)
+                    {
+                        concreteValues.Add(new InputEnumTypeStringValue(value.Name, (string)value.Value, value.Description));
+                    }
+                    break;
+                case InputTypeKind.Int32:
+                    foreach (var value in allowedValues)
+                    {
+                        concreteValues.Add(new InputEnumTypeIntegerValue(value.Name, (Int32)value.Value, value.Description));
+                    }
+                    break;
+                case InputTypeKind.Float32:
+                    foreach (var value in allowedValues)
+                    {
+                        concreteValues.Add(new InputEnumTypeFloatValue(value.Name, (float)value.Value, value.Description));
+                    }
+                    break;
+                default:
+                    throw new JsonException($"Unsupported enum value type: {valueType.Kind}");
+            }
+
+            return concreteValues;
+        }
+
+        //private static IReadOnlyList<InputEnumTypeValue> convertToConcreteValues<EnumType, ValueType>(IReadOnlyList<InputEnumTypeValue> allowedValues) where EnumType : InputEnumTypeValue
+        //{
+        //    var concreteValues = new List<InputEnumTypeValue>(allowedValues.Count);
+
+        //    foreach (var value in allowedValues)
+        //    {
+        //        concreteValues.Add(new EnumType(value.Name, (ValueType)value.Value, value.Description));
+        //    }
+
+        //    return concreteValues;
+        //}
     }
 }