--- a/src/AutoRest.CSharp/Common/Input/Utf8JsonReaderExtensions.cs
+++ b/src/AutoRest.CSharp/Common/Input/Utf8JsonReaderExtensions.cs
@@ -70,6 +70,46 @@ namespace AutoRest.CSharp.Common.Input
             return true;
         }
 
+        public static bool TryReadEnumValue(this ref Utf8JsonReader reader, string propertyName, ref object? value)
+        {
+            if (reader.TokenType != JsonTokenType.PropertyName)
+            {
+                throw new JsonException();
+            }
+
+            if (reader.GetString() != propertyName)
+            {
+                return false;
+            }
+
+            reader.Read();
+            switch (reader.TokenType)
+            {
+                case JsonTokenType.String:
+                    value = reader.GetString() ?? throw new JsonException();
+                    break;
+                case JsonTokenType.Number:
+                    if (reader.TryGetInt32(out int intValue))
+                    {
+                        value = intValue;
+                    }
+                    else if (reader.TryGetSingle(out float floatValue))
+                    {
+                        value = floatValue;
+                    }
+                    else
+                    {
+                        throw new JsonException($"Unsupported enum value type: {reader.TokenType}");
+                    }
+                    break;
+                default:
+                    throw new JsonException($"Unsupported enum value type: {reader.TokenType}");
+            }
+
+            reader.Read();
+            return true;
+        }
+
         public static bool TryReadPrimitiveType(this ref Utf8JsonReader reader, string propertyName, ref InputPrimitiveType? value)
         {
             if (reader.TokenType != JsonTokenType.PropertyName)