--- a/test/AutoRest.TestServer.Tests/Mgmt/TestProjects/MgmtExtensionResourceTests.cs
+++ b/test/AutoRest.TestServer.Tests/Mgmt/TestProjects/MgmtExtensionResourceTests.cs
@@ -1,5 +1,6 @@
-﻿using System.Linq;
-using System.Reflection;
+﻿using System;
+using System.Linq;
+using MgmtExpandResourceTypes.Models;
 using NUnit.Framework;
 
 namespace AutoRest.TestServer.Tests.Mgmt.TestProjects
@@ -26,6 +27,19 @@ namespace AutoRest.TestServer.Tests.Mgmt.TestProjects
             var classesToCheck = FindAllCollections().Concat(FindAllResources()).Append(FindExtensionClass());
             var classToCheck = classesToCheck.First(t => t.Name == className);
             Assert.AreEqual(exist, classToCheck.GetMethod(methodName) != null, $"can{(exist ? "not" : string.Empty)} find {className}.{methodName}");
+        }
+
+        [TestCase(typeof(MachineType), new string[] { "One", "Two", "Four" }, new int[] { 1, 2, 4 })]
+        [TestCase(typeof(StorageType), new string[] { "StandardLRS", "StandardZRS", "StandardGRS" }, new int[] { 1, 2, 3 })]
+        public void ValidateIntEnumValues(Type enumType, string[] expectedNames, int[] expectedValues)
+        {
+            var names = Enum.GetNames(enumType);
+            Assert.AreEqual(expectedNames, names);
+
+            for (int i = 0; i < expectedNames.Length; i++)
+            {
+                Assert.AreEqual(expectedValues[i], (int)Enum.Parse(enumType, expectedNames[i]));
+            }
         }
     }
 }