… test/TestProjects/MgmtExpandResourceTypes/Generated/CodeModel.yaml
… test/TestProjects/MgmtExpandResourceTypes/Generated/Models/MachineType.cs
… test/TestProjects/MgmtExpandResourceTypes/Generated/Models/StorageType.cs
… test/TestProjects/MgmtExpandResourceTypes/Generated/ZoneData.cs
… test/TestProjects/MgmtExpandResourceTypes/MgmtExpandResourceTypes.json
… test/TestProjects/Models-Cadl/Generated/Docs/ModelsInCadlClient.xml
… test/TestProjects/Models-Cadl/Generated/FixedIntEnum.cs
… test/TestProjects/Models-Cadl/Generated/RoundTripModel.Serialization.cs
… test/TestProjects/Models-Cadl/Generated/RoundTripModel.cs
… test/TestProjects/Models-Cadl/Generated/cadl.json
… test/TestProjects/Models-Cadl/Models-Cadl.cadl