- change `Value` property of `InputenumTypeValue` from `string` to `object`
- add new enum value typs: `InputEnumTypeStringValue`, `InputEnumTypeIntegerValue` and `InputEnumTypeFloatValue`
- change `CadlInputEnumTypeValueConverter` support parsing integer and float types of enum values
- change `CadlInputEnumTypeConverter` to specify more specific enum value types after initial parsing
- change serialization code for int enum, use `WriteNumberValue`
- add test case

resolve #2591

# Description
<i>Add your description here!</i>

# Checklist

To ensure a quick review and merge, please ensure:
- [ ] The PR has a understandable title and description explaining the _why_ and _what_.
- [ ] The PR is opened in draft if not ready for review yet.
   - If opened in draft, please allocate sufficient time (24 hours) after moving out of draft for review
- [ ] The branch is recent enough to not have merge conflicts upon creation.

# Ready to Land?
- [ ] Build is completely green
   - Submissions with test failures require tracking issue and approval of a CODEOWNER
- [ ] At least one +1 review by a CODEOWNER
- [ ] All -1 reviews are confirmed resolved by the reviewer 
   - Override/Marking reviews stale must be discussed with CODEOWNERS first