@archerzz 2022-10-31
This is a conservative solution to implement fixed int enum support. A few key points:
- It still generates serialization class. For int enum, we actually do not need serialization class because in .Net enum is int value. I keep it because if we remove serialization class, there will be many changes in existing generator codes.
- The parsing of incoming enum definition is a two-pass process.
  - In the first iteration, we parse the JSON enum definitions into `InputEnumTypeValue` which is a base class with `object Value` property.
  - In the second iteration, we will check the `ValueType` property of `InputEnumType` and then use that information to convert to more specific enum type values (e.g. `InputEnumTypeStringValue`, `InputEnumTypeIntegerValue` and `InputEnumTypeFloatValue`).
  - The process is not ideal, because I could not find a more elegant way to parse the enum value in one-pass.
    - We should not assume the sequence of properties of `InputEnumType`, so when parsing he `ValueType` and `AllowedValues` we could not assume the their sequence, thus we could not specify the exact value type when parsing `AllowedValues`.
    - I tried to get the raw bytes of `AllowedValues`, so that we could parse them after we parsed the `InputEnumType`. But I could not find a way to get the raw bytes. The stream API of `System.Text.Json` doesn't provide a way to fetch them.