@archerzz 2022-11-01
## Performance Comparison

I run the following benchmark codes.
```csharp
    [SimpleJob(RuntimeMoniker.Net461)]
    [SimpleJob(RuntimeMoniker.Net60)]
    [SimpleJob(RuntimeMoniker.NetCoreApp31)]
    public class IntEnumSerliazationBenchmark
    {
        [Params(1, 2, 4)]
        public int Value;

        [Benchmark]
        public FixedIntEnum UseDeserializationMethod()
        {
            return Value.ToFixedIntEnum();
        }

        [Benchmark]
        public FixedIntEnum UseSystemIntrinsicMethod()
        {
            if (Enum.IsDefined(typeof(FixedIntEnum), Value))
            {
                return (FixedIntEnum)Enum.ToObject(typeof(FixedIntEnum), Value);
            }
            else
            {
                throw new ArgumentOutOfRangeException(nameof(Value), Value, "Unknown FixedIntEnum value.");
            }
        }
    }
``

And the result is as follows.
```
|                   Method |                  Job |              Runtime | Value |       Mean |     Error |    StdDev |
|------------------------- |--------------------- |--------------------- |------ |-----------:|----------:|----------:|
| UseDeserializationMethod |             .NET 6.0 |             .NET 6.0 |     1 |   7.205 ns | 0.0479 ns | 0.0400 ns |
| UseSystemIntrinsicMethod |             .NET 6.0 |             .NET 6.0 |     1 | 164.200 ns | 0.5609 ns | 0.4683 ns |
| UseDeserializationMethod |        .NET Core 3.1 |        .NET Core 3.1 |     1 |  11.630 ns | 0.0524 ns | 0.0490 ns |
| UseSystemIntrinsicMethod |        .NET Core 3.1 |        .NET Core 3.1 |     1 | 170.961 ns | 1.1150 ns | 0.9884 ns |
| UseDeserializationMethod | .NET Framework 4.6.1 | .NET Framework 4.6.1 |     1 |   9.875 ns | 0.0994 ns | 0.0881 ns |
| UseSystemIntrinsicMethod | .NET Framework 4.6.1 | .NET Framework 4.6.1 |     1 | 194.658 ns | 1.4495 ns | 1.3559 ns |
| UseDeserializationMethod |             .NET 6.0 |             .NET 6.0 |     2 |  14.817 ns | 0.2768 ns | 0.2454 ns |
| UseSystemIntrinsicMethod |             .NET 6.0 |             .NET 6.0 |     2 | 159.392 ns | 1.2592 ns | 1.1778 ns |
| UseDeserializationMethod |        .NET Core 3.1 |        .NET Core 3.1 |     2 |  23.051 ns | 0.1183 ns | 0.1107 ns |
| UseSystemIntrinsicMethod |        .NET Core 3.1 |        .NET Core 3.1 |     2 | 169.208 ns | 1.0923 ns | 0.8528 ns |
| UseDeserializationMethod | .NET Framework 4.6.1 | .NET Framework 4.6.1 |     2 |  19.965 ns | 0.3222 ns | 0.3014 ns |
| UseSystemIntrinsicMethod | .NET Framework 4.6.1 | .NET Framework 4.6.1 |     2 | 195.447 ns | 1.3809 ns | 1.1531 ns |
| UseDeserializationMethod |             .NET 6.0 |             .NET 6.0 |     4 |  22.447 ns | 0.1357 ns | 0.1203 ns |
| UseSystemIntrinsicMethod |             .NET 6.0 |             .NET 6.0 |     4 | 159.490 ns | 0.5889 ns | 0.5220 ns |
| UseDeserializationMethod |        .NET Core 3.1 |        .NET Core 3.1 |     4 |  35.416 ns | 0.1923 ns | 0.1799 ns |
| UseSystemIntrinsicMethod |        .NET Core 3.1 |        .NET Core 3.1 |     4 | 174.281 ns | 1.2075 ns | 1.0704 ns |
| UseDeserializationMethod | .NET Framework 4.6.1 | .NET Framework 4.6.1 |     4 |  29.862 ns | 0.1925 ns | 0.1503 ns |
| UseSystemIntrinsicMethod | .NET Framework 4.6.1 | .NET Framework 4.6.1 |     4 | 200.741 ns | 1.3288 ns | 1.2429 ns |
```

Deserialization method in the serialization class is way faster than system intrinsic methods.