@archerzz 2022-11-01
list of changes:
- add assignment of int value to enum values, like `One = 1,`
- name the serialization method after `ToSerial{Type}`
  - for existing enum types, the serialization method name won't change
- add test case for mgmt plane