@archerzz 2022-11-07
## Benchmark for optimized string deserialization

I changed the deserialization codes in:

- Use `StringComparison.OrdinalIgnoreCase` for string comparision
- Use `String.AsSpan()` for `JsonElement.ValueEquals()`

The updated codes are as follows:

```csharp
    internal static partial class FixedStringEnumExtensions
    {
        private const string One = "1";
        private const string Two = "2";
        private const string Four = "4";

        public static string ToSerialString(this FixedStringEnum value) => value switch
        {
            FixedStringEnum.One => "1",
            FixedStringEnum.Two => "2",
            FixedStringEnum.Four => "4",
            _ => throw new ArgumentOutOfRangeException(nameof(value), value, "Unknown FixedStringEnum value.")
        };

        public static FixedStringEnum ToFixedStringEnum(this string value)
        {
            if (string.Equals(value, "1", StringComparison.OrdinalIgnoreCase)) return FixedStringEnum.One;
            if (string.Equals(value, "2", StringComparison.OrdinalIgnoreCase)) return FixedStringEnum.Two;
            if (string.Equals(value, "4", StringComparison.OrdinalIgnoreCase)) return FixedStringEnum.Four;
            throw new ArgumentOutOfRangeException(nameof(value), value, "Unknown FixedStringEnum value.");
        }

        public static FixedStringEnum ToFixedStringEnum(this JsonElement value)
        {
            if (value.ValueEquals("1".AsSpan()))
                return FixedStringEnum.One;
            if (value.ValueEquals("2".AsSpan()))
                return FixedStringEnum.Two;
            if (value.ValueEquals("4".AsSpan()))
                return FixedStringEnum.Four;
            throw new ArgumentOutOfRangeException(nameof(value), value, "Unknown FixedStringEnum value.");
        }

        public static FixedStringEnum ToFixedStringEnumWithConst(this JsonElement value)
        {
            if (value.ValueEquals(One.AsSpan()))
                return FixedStringEnum.One;
            if (value.ValueEquals(Two.AsSpan()))
                return FixedStringEnum.Two;
            if (value.ValueEquals(Four.AsSpan()))
                return FixedStringEnum.Four;
            throw new ArgumentOutOfRangeException(nameof(value), value, "Unknown FixedStringEnum value.");
        }
    }
```


### Performance

TL;DR

- String comparison is much faster, and it's even faster than `JsonElement.ValueEquals()`
- There is little difference between new/old versions 

| Method                    | Job                  | Runtime              | JsonValue | Before    | After     | Difference | Percentage |
|---------------------------|----------------------|----------------------|-----------|-----------|-----------|------------|------------|
| UseDeserializationMethod  | .NET 6.0             | .NET 6.0             | 1         | 33.16 ns  | 11.28 ns  | -21.9      | -66.16%    |
| UseJsonElementMethod      | .NET 6.0             | .NET 6.0             | 1         | 27.32 ns  | 28.22 ns  | 0.9        | 3.30%      |
| UseJsonElementConstMethod | .NET 6.0             | .NET 6.0             | 1         | 27.12 ns  | 27.79 ns  | 0.6        | 2.21%      |
| UseDeserializationMethod  | .NET Core 3.1        | .NET Core 3.1        | 1         | 53.20 ns  | 10.84 ns  | -42.4      | -79.70%    |
| UseJsonElementMethod      | .NET Core 3.1        | .NET Core 3.1        | 1         | 37.64 ns  | 37.12 ns  | -0.5       | -1.33%     |
| UseJsonElementConstMethod | .NET Core 3.1        | .NET Core 3.1        | 1         | 37.19 ns  | 38.16 ns  | 1          | 2.70%      |
| UseDeserializationMethod  | .NET Framework 4.6.1 | .NET Framework 4.6.1 | 1         | 77.83 ns  | 21.37 ns  | -56.5      | -72.62%    |
| UseJsonElementMethod      | .NET Framework 4.6.1 | .NET Framework 4.6.1 | 1         | 101.07 ns | 109.74 ns | 8.7        | 8.61%      |
| UseJsonElementConstMethod | .NET Framework 4.6.1 | .NET Framework 4.6.1 | 1         | 99.96 ns  | 110.05 ns | 10.1       | 10.11%     |
| UseDeserializationMethod  | .NET 6.0             | .NET 6.0             | 2         | 70.82 ns  | 16.59 ns  | -54.3      | -76.69%    |
| UseJsonElementMethod      | .NET 6.0             | .NET 6.0             | 2         | 56.89 ns  | 52.38 ns  | -4.5       | -7.92%     |
| UseJsonElementConstMethod | .NET 6.0             | .NET 6.0             | 2         | 51.93 ns  | 52.07 ns  | 0.1        | 0.19%      |
| UseDeserializationMethod  | .NET Core 3.1        | .NET Core 3.1        | 2         | 97.57 ns  | 19.28 ns  | -78.3      | -80.31%    |
| UseJsonElementMethod      | .NET Core 3.1        | .NET Core 3.1        | 2         | 71.31 ns  | 72.60 ns  | 1.3        | 1.82%      |
| UseJsonElementConstMethod | .NET Core 3.1        | .NET Core 3.1        | 2         | 73.84 ns  | 72.78 ns  | -1.1       | -1.49%     |
| UseDeserializationMethod  | .NET Framework 4.6.1 | .NET Framework 4.6.1 | 2         | 153.37 ns | 29.97 ns  | -123.4     | -80.50%    |
| UseJsonElementMethod      | .NET Framework 4.6.1 | .NET Framework 4.6.1 | 2         | 201.13 ns | 209.65 ns | 8.5        | 4.23%      |
| UseJsonElementConstMethod | .NET Framework 4.6.1 | .NET Framework 4.6.1 | 2         | 195.09 ns | 209.63 ns | 14.6       | 7.49%      |
| UseDeserializationMethod  | .NET 6.0             | .NET 6.0             | 4         | 107.37 ns | 22.00 ns  | -85.3      | -79.50%    |
| UseJsonElementMethod      | .NET 6.0             | .NET 6.0             | 4         | 80.12 ns  | 78.21 ns  | -1.9       | -2.37%     |
| UseJsonElementConstMethod | .NET 6.0             | .NET 6.0             | 4         | 79.11 ns  | 76.96 ns  | -2.2       | -2.78%     |
| UseDeserializationMethod  | .NET Core 3.1        | .NET Core 3.1        | 4         | 146.61 ns | 25.66 ns  | -121       | -82.54%    |
| UseJsonElementMethod      | .NET Core 3.1        | .NET Core 3.1        | 4         | 107.91 ns | 107.66 ns | -0.3       | -0.28%     |
| UseJsonElementConstMethod | .NET Core 3.1        | .NET Core 3.1        | 4         | 107.68 ns | 105.87 ns | -1.8       | -1.67%     |
| UseDeserializationMethod  | .NET Framework 4.6.1 | .NET Framework 4.6.1 | 4         | 210.75 ns | 37.75 ns  | -173       | -82.11%    |
| UseJsonElementMethod      | .NET Framework 4.6.1 | .NET Framework 4.6.1 | 4         | 289.61 ns | 314.55 ns | 24.9       | 8.60%      |
| UseJsonElementConstMethod | .NET Framework 4.6.1 | .NET Framework 4.6.1 | 4         | 287.90 ns | 314.32 ns | 26.4       | 9.17%      |